ArrayIndexOutOfBoundException for DictionaryExclude and noInvertedIndex
