+++ core/src/main/java/org/apache/carbondata/core/datastore/columnar/BlockIndexerStorageForNoInvertedIndexForShort.java
@@ -36,22 +36,39 @@ public class BlockIndexerStorageForNoInvertedIndexForShort implements IndexStora
   private byte[] min;
   private byte[] max;
 
-  public BlockIndexerStorageForNoInvertedIndexForShort(byte[][] keyBlockInput) {
+  public BlockIndexerStorageForNoInvertedIndexForShort(byte[][] keyBlockInput,boolean isNoDictonary)
+  {
     this.keyBlock = keyBlockInput;
     min = keyBlock[0];
     max = keyBlock[0];
     totalSize += keyBlock[0].length;
     int minCompare = 0;
     int maxCompare = 0;
-    for (int i = 1; i < keyBlock.length; i++) {
-      totalSize += keyBlock[i].length;
-      minCompare = ByteUtil.compare(min, keyBlock[i]);
-      maxCompare = ByteUtil.compare(max, keyBlock[i]);
-      if (minCompare > 0) {
-        min = keyBlock[i];
+    if (!isNoDictonary) {
+      for (int i = 1; i < keyBlock.length; i++) {
+        totalSize += keyBlock[i].length;
+        minCompare = ByteUtil.compare(min, keyBlock[i]);
+        maxCompare = ByteUtil.compare(max, keyBlock[i]);
+        if (minCompare > 0) {
+          min = keyBlock[i];
+        }
+        if (maxCompare < 0) {
+          max = keyBlock[i];
+        }
       }
-      if (maxCompare < 0) {
-        max = keyBlock[i];
+    } else {
+      for (int i = 1; i < keyBlock.length; i++) {
+        totalSize += keyBlock[i].length;
+        minCompare = ByteUtil.UnsafeComparer.INSTANCE
+            .compareTo(min, 2, min.length - 2, keyBlock[i], 2, keyBlock[i].length - 2);
+        maxCompare = ByteUtil.UnsafeComparer.INSTANCE
+            .compareTo(max, 2, max.length - 2, keyBlock[i], 2, keyBlock[i].length - 2);
+        if (minCompare > 0) {
+          min = keyBlock[i];
+        }
+        if (maxCompare < 0) {
+          max = keyBlock[i];
+        }
       }
     }
   }
